Refactoring, rearranging.
